Fix build errors in `SearchViewModelTests.swift`
Some of the references to `Polygon` were not qualified when they needed to be, so they were resolving to the `Polygon` type in the Application Services framework. While I was at it, I made the geometries computed so that they don't stick around longer than necessary. We've also seen problems with re-using geometries.